Add a function to fix autofill problem
Create a function to check input value and mimic client interaction by
setting its value. this is due to input not registerng autofill after
user going back to the login page by clicking back button on login error
page. This causes a bug where input.value is not empty but the input
acts as if it is empty. focusing the input doesn't seem to register the
input and the only way for the input to register its value is by
triggering input change.